願望清單v1

+++ backend/src/main/java/org/ezlearn/model/PurchasedCourses.java
@@ -20,7 +20,7 @@ public class PurchasedCourses {
 	private PurchasedCoursesId purchasedCoursesId;
 	
 	private String courseReview;
-	private Double courseRate;
+	private Integer courseRate;
 	
 
 	public String getCourseReview() {
@@ -29,10 +29,10 @@ public class PurchasedCourses {
 	public void setCourseReview(String courseReview) {
 		this.courseReview = courseReview;
 	}
-	public Double getCourseRate() {
+	public Integer getCourseRate() {
 		return courseRate;
 	}
-	public void setCourseRate(Double courseRate) {
+	public void setCourseRate(Integer courseRate) {
 		this.courseRate = courseRate;
 	}
 	public PurchasedCoursesId getPurchasedCoursesId() {
